Move more reduction tests
