+++ python/cudf/cudf/tests/groupby/test_agg.py
@@ -1,4 +1,6 @@
 # Copyright (c) 2023-2025, NVIDIA CORPORATION.
+import decimal
+
 import numpy as np
 import pandas as pd
 import pytest
@@ -193,3 +195,62 @@ def test_series_groupby_agg(groupby_reduction_methods):
     sg = s.groupby(s // 2).agg(groupby_reduction_methods)
     gg = g.groupby(g // 2).agg(groupby_reduction_methods)
     assert_groupby_results_equal(sg, gg)
+
+
+def test_groupby_agg_decimal(groupby_reduction_methods, request):
+    request.applymarker(
+        pytest.mark.xfail(
+            groupby_reduction_methods in ["prod", "mean"],
+            raises=pd.errors.DataError,
+            reason=f"{groupby_reduction_methods} not supported with Decimals in pandas",
+        )
+    )
+    rng = np.random.default_rng(seed=0)
+    num_groups = 4
+    nelem_per_group = 10
+    # The number of digits after the decimal to use.
+    decimal_digits = 2
+    # The number of digits before the decimal to use.
+    whole_digits = 2
+
+    scale = 10**whole_digits
+    nelem = num_groups * nelem_per_group
+
+    # The unique is necessary because otherwise if there are duplicates idxmin
+    # and idxmax may return different results than pandas (see
+    # https://github.com/rapidsai/cudf/issues/7756). This is not relevant to
+    # the current version of the test, because idxmin and idxmax simply don't
+    # work with pandas Series composed of Decimal objects (see
+    # https://github.com/pandas-dev/pandas/issues/40685). However, if that is
+    # ever enabled, then this issue will crop up again so we may as well have
+    # it fixed now.
+    x = np.unique((rng.random(nelem) * scale).round(decimal_digits))
+    y = np.unique((rng.random(nelem) * scale).round(decimal_digits))
+
+    if x.size < y.size:
+        total_elements = x.size
+        y = y[: x.size]
+    else:
+        total_elements = y.size
+        x = x[: y.size]
+
+    # Note that this filtering can lead to one group with fewer elements, but
+    # that shouldn't be a problem and is probably useful to test.
+    idx_col = np.tile(np.arange(num_groups), nelem_per_group)[:total_elements]
+
+    decimal_x = pd.Series([decimal.Decimal(str(d)) for d in x])
+    decimal_y = pd.Series([decimal.Decimal(str(d)) for d in y])
+
+    pdf = pd.DataFrame({"idx": idx_col, "x": decimal_x, "y": decimal_y})
+    gdf = cudf.DataFrame(
+        {
+            "idx": idx_col,
+            "x": cudf.Series(decimal_x),
+            "y": cudf.Series(decimal_y),
+        }
+    )
+
+    expect_df = pdf.groupby("idx", sort=True).agg(groupby_reduction_methods)
+    got_df = gdf.groupby("idx", sort=True).agg(groupby_reduction_methods)
+    assert_eq(expect_df["x"], got_df["x"], check_dtype=False)
+    assert_eq(expect_df["y"], got_df["y"], check_dtype=False)
