+++ python/cudf/cudf/tests/groupby/test_reductions.py
@@ -1,8 +1,8 @@
 # Copyright (c) 2025, NVIDIA CORPORATION.
 
-
 import numpy as np
 import pandas as pd
+import pytest
 
 import cudf
 from cudf.tests.groupby.testing import assert_groupby_results_equal
@@ -90,3 +90,133 @@ def test_series_groupby(groupby_reduction_methods):
     sa = getattr(sg, groupby_reduction_methods)()
     ga = getattr(gg, groupby_reduction_methods)()
     assert_groupby_results_equal(sa, ga)
+
+
+def test_groupby_level_zero(groupby_reduction_methods, request):
+    request.applymarker(
+        pytest.mark.xfail(
+            groupby_reduction_methods in ["idxmin", "idxmax"],
+            reason="gather needed for idxmin/idxmax",
+        )
+    )
+    pdf = pd.DataFrame({"x": [1, 2, 3]}, index=[2, 5, 5])
+    gdf = cudf.DataFrame.from_pandas(pdf)
+    pdg = pdf.groupby(level=0)
+    gdg = gdf.groupby(level=0)
+    pdresult = getattr(pdg, groupby_reduction_methods)()
+    gdresult = getattr(gdg, groupby_reduction_methods)()
+    assert_groupby_results_equal(
+        pdresult,
+        gdresult,
+    )
+
+
+def test_groupby_series_level_zero(groupby_reduction_methods, request):
+    request.applymarker(
+        pytest.mark.xfail(
+            groupby_reduction_methods in ["idxmin", "idxmax"],
+            reason="gather needed for idxmin/idxmax",
+        )
+    )
+    pdf = pd.Series([1, 2, 3], index=[2, 5, 5])
+    gdf = cudf.Series.from_pandas(pdf)
+    pdg = pdf.groupby(level=0)
+    gdg = gdf.groupby(level=0)
+    pdresult = getattr(pdg, groupby_reduction_methods)()
+    gdresult = getattr(gdg, groupby_reduction_methods)()
+    assert_groupby_results_equal(pdresult, gdresult)
+
+
+def test_groupby_column_name():
+    pdf = pd.DataFrame({"xx": [1.0, 2.0, 3.0], "yy": [1, 2, 3]})
+    gdf = cudf.DataFrame.from_pandas(pdf)
+    g = gdf.groupby("yy")
+    p = pdf.groupby("yy")
+    gxx = g["xx"].sum()
+    pxx = p["xx"].sum()
+    assert_groupby_results_equal(pxx, gxx)
+
+    gxx = g["xx"].count()
+    pxx = p["xx"].count()
+    assert_groupby_results_equal(pxx, gxx, check_dtype=False)
+
+    gxx = g["xx"].min()
+    pxx = p["xx"].min()
+    assert_groupby_results_equal(pxx, gxx)
+
+    gxx = g["xx"].max()
+    pxx = p["xx"].max()
+    assert_groupby_results_equal(pxx, gxx)
+
+    gxx = g["xx"].idxmin()
+    pxx = p["xx"].idxmin()
+    assert_groupby_results_equal(pxx, gxx, check_dtype=False)
+
+    gxx = g["xx"].idxmax()
+    pxx = p["xx"].idxmax()
+    assert_groupby_results_equal(pxx, gxx, check_dtype=False)
+
+    gxx = g["xx"].mean()
+    pxx = p["xx"].mean()
+    assert_groupby_results_equal(pxx, gxx)
+
+
+def test_groupby_column_numeral():
+    pdf = pd.DataFrame({0: [1.0, 2.0, 3.0], 1: [1, 2, 3]})
+    gdf = cudf.DataFrame.from_pandas(pdf)
+    p = pdf.groupby(1)
+    g = gdf.groupby(1)
+    pxx = p[0].sum()
+    gxx = g[0].sum()
+    assert_groupby_results_equal(pxx, gxx)
+
+    pdf = pd.DataFrame({0.5: [1.0, 2.0, 3.0], 1.5: [1, 2, 3]})
+    gdf = cudf.DataFrame.from_pandas(pdf)
+    p = pdf.groupby(1.5)
+    g = gdf.groupby(1.5)
+    pxx = p[0.5].sum()
+    gxx = g[0.5].sum()
+    assert_groupby_results_equal(pxx, gxx)
+
+
+@pytest.mark.parametrize(
+    "series",
+    [
+        [0, 1, 0],
+        [1, 1, 1],
+        [0, 1, 1],
+        [1, 2, 3],
+        [4, 3, 2],
+        [0, 2, 0],
+        pd.Series([0, 2, 0]),
+        pd.Series([0, 2, 0], index=[0, 2, 1]),
+    ],
+)
+def test_groupby_external_series(series):
+    pdf = pd.DataFrame({"x": [1.0, 2.0, 3.0], "y": [1, 2, 1]})
+    gdf = cudf.DataFrame.from_pandas(pdf)
+    pxx = pdf.groupby(pd.Series(series)).x.sum()
+    gxx = gdf.groupby(cudf.Series(series)).x.sum()
+    assert_groupby_results_equal(pxx, gxx)
+
+
+@pytest.mark.parametrize("series", [[0.0, 1.0], [1.0, 1.0, 1.0, 1.0]])
+def test_groupby_external_series_incorrect_length(series):
+    pdf = pd.DataFrame({"x": [1.0, 2.0, 3.0], "y": [1, 2, 1]})
+    gdf = cudf.DataFrame.from_pandas(pdf)
+    pxx = pdf.groupby(pd.Series(series)).x.sum()
+    gxx = gdf.groupby(cudf.Series(series)).x.sum()
+    assert_groupby_results_equal(pxx, gxx)
+
+
+@pytest.mark.parametrize(
+    "level", [0, 1, "a", "b", [0, 1], ["a", "b"], ["a", 1], -1, [-1, -2]]
+)
+def test_groupby_levels(level):
+    idx = pd.MultiIndex.from_tuples([(1, 1), (1, 2), (2, 2)], names=("a", "b"))
+    pdf = pd.DataFrame({"c": [1, 2, 3], "d": [2, 3, 4]}, index=idx)
+    gdf = cudf.from_pandas(pdf)
+    assert_groupby_results_equal(
+        pdf.groupby(level=level).sum(),
+        gdf.groupby(level=level).sum(),
+    )
