+++ python/cudf/cudf/tests/test_groupby.py
@@ -6,7 +6,6 @@ import itertools
 import operator
 import string
 import textwrap
-from decimal import Decimal
 from functools import partial
 
 import numpy as np
@@ -692,220 +691,6 @@ def test_groupby_apply_return_reindexed_series(as_index):
     expect = pdf.groupby("key", as_index=as_index).apply(pdf_func, **kwargs)
     got = df.groupby("key", as_index=as_index).apply(gdf_func, **kwargs)
     assert_groupby_results_equal(expect, got)
-
-
-@pytest.mark.parametrize("num_groups", [2, 20])
-@pytest.mark.parametrize("nelem_per_group", [1, 10])
-@pytest.mark.parametrize(
-    "func",
-    ["min", "max", "count", "sum"],
-    # TODO: Replace the above line with the one below once
-    # https://github.com/pandas-dev/pandas/issues/40685 is resolved.
-    # "func", ["min", "max", "idxmin", "idxmax", "count", "sum"],
-)
-def test_groupby_agg_decimal(num_groups, nelem_per_group, func):
-    rng = np.random.default_rng(seed=0)
-    # The number of digits after the decimal to use.
-    decimal_digits = 2
-    # The number of digits before the decimal to use.
-    whole_digits = 2
-
-    scale = 10**whole_digits
-    nelem = num_groups * nelem_per_group
-
-    # The unique is necessary because otherwise if there are duplicates idxmin
-    # and idxmax may return different results than pandas (see
-    # https://github.com/rapidsai/cudf/issues/7756). This is not relevant to
-    # the current version of the test, because idxmin and idxmax simply don't
-    # work with pandas Series composed of Decimal objects (see
-    # https://github.com/pandas-dev/pandas/issues/40685). However, if that is
-    # ever enabled, then this issue will crop up again so we may as well have
-    # it fixed now.
-    x = np.unique((rng.random(nelem) * scale).round(decimal_digits))
-    y = np.unique((rng.random(nelem) * scale).round(decimal_digits))
-
-    if x.size < y.size:
-        total_elements = x.size
-        y = y[: x.size]
-    else:
-        total_elements = y.size
-        x = x[: y.size]
-
-    # Note that this filtering can lead to one group with fewer elements, but
-    # that shouldn't be a problem and is probably useful to test.
-    idx_col = np.tile(np.arange(num_groups), nelem_per_group)[:total_elements]
-
-    decimal_x = pd.Series([Decimal(str(d)) for d in x])
-    decimal_y = pd.Series([Decimal(str(d)) for d in y])
-
-    pdf = pd.DataFrame({"idx": idx_col, "x": decimal_x, "y": decimal_y})
-    gdf = DataFrame(
-        {
-            "idx": idx_col,
-            "x": cudf.Series(decimal_x),
-            "y": cudf.Series(decimal_y),
-        }
-    )
-
-    expect_df = pdf.groupby("idx", sort=True).agg(func)
-    got_df = gdf.groupby("idx", sort=True).agg(func)
-    assert_eq(expect_df["x"], got_df["x"], check_dtype=False)
-    assert_eq(expect_df["y"], got_df["y"], check_dtype=False)
-
-
-@pytest.mark.parametrize(
-    "agg",
-    [
-        "min",
-        "max",
-        "count",
-        "sum",
-        "prod",
-        "mean",
-        pytest.param(
-            "idxmin",
-            marks=pytest.mark.xfail(reason="gather needed for idxmin"),
-        ),
-        pytest.param(
-            "idxmax",
-            marks=pytest.mark.xfail(reason="gather needed for idxmax"),
-        ),
-    ],
-)
-def test_groupby_level_zero(agg):
-    pdf = pd.DataFrame({"x": [1, 2, 3]}, index=[2, 5, 5])
-    gdf = DataFrame.from_pandas(pdf)
-    pdg = pdf.groupby(level=0)
-    gdg = gdf.groupby(level=0)
-    pdresult = getattr(pdg, agg)()
-    gdresult = getattr(gdg, agg)()
-    check_dtype = agg not in _index_type_aggs
-    assert_groupby_results_equal(pdresult, gdresult, check_dtype=check_dtype)
-
-
-@pytest.mark.parametrize(
-    "agg",
-    [
-        "min",
-        "max",
-        "count",
-        "sum",
-        "prod",
-        "mean",
-        pytest.param(
-            "idxmin",
-            marks=pytest.mark.xfail(reason="gather needed for idxmin"),
-        ),
-        pytest.param(
-            "idxmax",
-            marks=pytest.mark.xfail(reason="gather needed for idxmax"),
-        ),
-    ],
-)
-def test_groupby_series_level_zero(agg):
-    pdf = pd.Series([1, 2, 3], index=[2, 5, 5])
-    gdf = Series.from_pandas(pdf)
-    pdg = pdf.groupby(level=0)
-    gdg = gdf.groupby(level=0)
-    pdresult = getattr(pdg, agg)()
-    gdresult = getattr(gdg, agg)()
-    check_dtype = agg not in _index_type_aggs
-    assert_groupby_results_equal(pdresult, gdresult, check_dtype=check_dtype)
-
-
-def test_groupby_column_name():
-    pdf = pd.DataFrame({"xx": [1.0, 2.0, 3.0], "yy": [1, 2, 3]})
-    gdf = DataFrame.from_pandas(pdf)
-    g = gdf.groupby("yy")
-    p = pdf.groupby("yy")
-    gxx = g["xx"].sum()
-    pxx = p["xx"].sum()
-    assert_groupby_results_equal(pxx, gxx)
-
-    gxx = g["xx"].count()
-    pxx = p["xx"].count()
-    assert_groupby_results_equal(pxx, gxx, check_dtype=False)
-
-    gxx = g["xx"].min()
-    pxx = p["xx"].min()
-    assert_groupby_results_equal(pxx, gxx)
-
-    gxx = g["xx"].max()
-    pxx = p["xx"].max()
-    assert_groupby_results_equal(pxx, gxx)
-
-    gxx = g["xx"].idxmin()
-    pxx = p["xx"].idxmin()
-    assert_groupby_results_equal(pxx, gxx, check_dtype=False)
-
-    gxx = g["xx"].idxmax()
-    pxx = p["xx"].idxmax()
-    assert_groupby_results_equal(pxx, gxx, check_dtype=False)
-
-    gxx = g["xx"].mean()
-    pxx = p["xx"].mean()
-    assert_groupby_results_equal(pxx, gxx)
-
-
-def test_groupby_column_numeral():
-    pdf = pd.DataFrame({0: [1.0, 2.0, 3.0], 1: [1, 2, 3]})
-    gdf = DataFrame.from_pandas(pdf)
-    p = pdf.groupby(1)
-    g = gdf.groupby(1)
-    pxx = p[0].sum()
-    gxx = g[0].sum()
-    assert_groupby_results_equal(pxx, gxx)
-
-    pdf = pd.DataFrame({0.5: [1.0, 2.0, 3.0], 1.5: [1, 2, 3]})
-    gdf = DataFrame.from_pandas(pdf)
-    p = pdf.groupby(1.5)
-    g = gdf.groupby(1.5)
-    pxx = p[0.5].sum()
-    gxx = g[0.5].sum()
-    assert_groupby_results_equal(pxx, gxx)
-
-
-@pytest.mark.parametrize(
-    "series",
-    [
-        [0, 1, 0],
-        [1, 1, 1],
-        [0, 1, 1],
-        [1, 2, 3],
-        [4, 3, 2],
-        [0, 2, 0],
-        pd.Series([0, 2, 0]),
-        pd.Series([0, 2, 0], index=[0, 2, 1]),
-    ],
-)
-def test_groupby_external_series(series):
-    pdf = pd.DataFrame({"x": [1.0, 2.0, 3.0], "y": [1, 2, 1]})
-    gdf = DataFrame.from_pandas(pdf)
-    pxx = pdf.groupby(pd.Series(series)).x.sum()
-    gxx = gdf.groupby(cudf.Series(series)).x.sum()
-    assert_groupby_results_equal(pxx, gxx)
-
-
-@pytest.mark.parametrize("series", [[0.0, 1.0], [1.0, 1.0, 1.0, 1.0]])
-def test_groupby_external_series_incorrect_length(series):
-    pdf = pd.DataFrame({"x": [1.0, 2.0, 3.0], "y": [1, 2, 1]})
-    gdf = DataFrame.from_pandas(pdf)
-    pxx = pdf.groupby(pd.Series(series)).x.sum()
-    gxx = gdf.groupby(cudf.Series(series)).x.sum()
-    assert_groupby_results_equal(pxx, gxx)
-
-
-@pytest.mark.parametrize(
-    "level", [0, 1, "a", "b", [0, 1], ["a", "b"], ["a", 1], -1, [-1, -2]]
-)
-def test_groupby_levels(level):
-    idx = pd.MultiIndex.from_tuples([(1, 1), (1, 2), (2, 2)], names=("a", "b"))
-    pdf = pd.DataFrame({"c": [1, 2, 3], "d": [2, 3, 4]}, index=idx)
-    gdf = cudf.from_pandas(pdf)
-    assert_groupby_results_equal(
-        pdf.groupby(level=level).sum(),
-        gdf.groupby(level=level).sum(),
-    )
 
 
 def test_advanced_groupby_levels():
